requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0
